# shared/openapi.json
{
  "components": {
    "schemas": {
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "LegOption": {
        "description": "Details of an individual journey leg, transit or walking.",
        "properties": {
          "arrival": {
            "title": "Arrival",
            "type": "string"
          },
          "connection_waiting_mins": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "default": 0,
            "title": "Connection Waiting Mins"
          },
          "departure": {
            "title": "Departure",
            "type": "string"
          },
          "duration_mins": {
            "title": "Duration Mins",
            "type": "integer"
          },
          "instruction": {
            "title": "Instruction",
            "type": "string"
          },
          "line": {
            "title": "Line",
            "type": "string"
          },
          "mode": {
            "title": "Mode",
            "type": "string"
          },
          "stops": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stops"
          }
        },
        "required": [
          "mode",
          "line",
          "duration_mins",
          "departure",
          "arrival",
          "instruction"
        ],
        "title": "LegOption",
        "type": "object"
      },
...
      "RouteFeature": {
        "description": "Color-coded sensory/efficiency feature badge for the route card.",
        "properties": {
          "color": {
            "title": "Color",
            "type": "string"
          },
          "icon": {
            "title": "Icon",
            "type": "string"
          },
          "label": {
            "title": "Label",
            "type": "string"
          },
          "type": {
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "type",
          "label",
          "icon",
          "color"
        ],
        "title": "RouteFeature",
        "type": "object"
      },
      "RouteOption": {
        "description": "A single route suggestion returned by ``GET /routes/``.\n\nField names mirror the contract the React Native client consumes exactly,\nincluding the camelCase ``subName`` alongside the snake_case sensory fields.\nThis model is the single source of truth for the route contract: the\nfrontend's TS types are generated from it (see ``shared/`` and\n``scripts/export_openapi.py``). Annotations are kept precise so the generated\ntypes are precise, but JSON serialization is unchanged (``SensoryLevel`` still\nserializes as the integers 1/2/3).",
        "properties": {
          "crowds": {
            "enum": [
              1,
              2,
              3,
              4
            ],
            "title": "Crowds",
            "type": "integer"
          },
          "description": {
            "title": "Description",
            "type": "string"
          },
          "duration": {
            "title": "Duration",
            "type": "integer"
          },
          "features": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/RouteFeature"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Features"
          },
          "heat": {
            "enum": [
              1,
              2,
              3,
              4
            ],
            "title": "Heat",
            "type": "integer"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "legs": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/LegOption"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Legs"
          },
          "light": {
            "enum": [
              1,
              2,
              3,
              4
            ],
            "title": "Light",
            "type": "integer"
          },
          "match_percentage": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Match Percentage"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "noise": {
            "enum": [
              1,
              2,
              3,
              4
            ],
            "title": "Noise",
            "type": "integer"
          },
          "price": {
            "title": "Price",
            "type": "number"
          },
          "sensory_description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sensory Description"
          },
          "sensory_score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sensory Score"
          },
          "smell": {
            "enum": [
              1,
              2,
              3,
              4
            ],
            "title": "Smell",
            "type": "integer"
          },
          "subName": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Subname"
          },
          "type": {
            "enum": [
              "best",
              "quickest",
              "suggested"
            ],
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "duration",
          "price",
          "noise",
          "crowds",
          "heat",
          "light",
          "smell",
          "description",
          "type"
        ],
        "title": "RouteOption",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
      "WarningItemSchema": {
        "description": "Client-facing live warning item schema, filtered by user sensory sensitivities.",
        "properties": {
          "desc": {
            "title": "Desc",
            "type": "string"
          },
          "icon": {
            "title": "Icon",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "severity": {
            "enum": [
              "high",
              "medium",
              "info"
            ],
            "title": "Severity",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "id",
          "title",
          "desc",
          "severity",
          "icon"
        ],
        "title": "WarningItemSchema",
        "type": "object"
      }
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/auth/login"
          }
        },
        "type": "oauth2"
      }
    }
  },
  "info": {
    "title": "Calm Travel API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/routes/": {
      "get": {
        "description": "Return route suggestions, optionally personalized to a user's sensitivities.",
        "operationId": "get_routes_routes__get",
        "parameters": [
          {
            "in": "query",
            "name": "start",
            "required": true,
            "schema": {
              "title": "Start",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "end",
            "required": true,
            "schema": {
              "title": "End",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "username",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Username"
            }
          },
          {
            "in": "query",
            "name": "walking_speed",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "default": "slow",
              "title": "Walking Speed"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/RouteOption"
                  },
                  "title": "Response Get Routes Routes  Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Get Routes",
        "tags": [
          "routes"
        ]
      }
    },
    "/routes/warnings": {
      "get": {
        "description": "Return live warnings tailored to the user's sensory sensitivities.",
        "operationId": "get_routes_warnings_routes_warnings_get",
        "parameters": [
          {
            "in": "query",
            "name": "username",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Username"
            }
          },
          {
            "in": "query",
            "name": "generic",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "default": false,
              "title": "Generic"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/WarningItemSchema"
                  },
                  "title": "Response Get Routes Warnings Routes Warnings Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Get Routes Warnings",
        "tags": [
          "routes"
        ]
      }
    },
...
  }
}